requirements.txt
tqdm
pytorch
pandas
numpy
pyarrow